Make object mapper column matching more flexible (#211)
This makes the column name to C# name matching more flexible. Symbol
characters are now ignored when matching, so a Kudu column `First_Name`
would match a C# property named `FirstName` (and vice versa).

When there are multiple matches the resolution order is,
1) An exact match
2) The first case insensitive match
3) The first match (ignoring symbols, whitespace, etc.)

This extended matching is not available on .NET Standard 2.0 due to a
missing overload of StringComparer.Create().